requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy==2.2.6
tensorflow==2.20.0
keras==3.10.0
